update5

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,6 @@
-monthly_income = int(input("Enter your monthly income: "))
-total_monthly_expenses = int(input("Enter your total monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 
-# Ensure monthly_savings is calculated as monthly_income - total_monthly_expenses
 monthly_savings = monthly_income - total_monthly_expenses
 project_savings = monthly_savings * 12 + (monthly_savings * 12 * 0.05)
